requirements.txt
# Core dependencies
jsonschema>=4.0.0
PyYAML>=6.0
pydantic>=2.0  # For advanced data validation (optional)

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Optional - for enhanced features
# python-dotenv>=1.0.0  # For environment variables
# structlog>=23.0.0    # For advanced logging